feat(29-02): migrate tests + example to nt_hash types, delete legacy hash tests
- test_resource.c: ~100 call sites migrated to nt_hash32_str/nt_hash64_str
- Deleted 3 hash tests (test_hash_known_value, test_hash_null, test_hash_empty) now in test_hash.c
- test_pack_format.c: updated NtAssetEntry size to 24 and field offsets for new layout
- textured_quad example: migrated to nt_hash32_t/nt_hash64_t for pack/resource IDs
- Fixed clang-tidy implicit widening in slot_map_find/slot_map_insert

+++ engine/resource/nt_resource.c
@@ -57,7 +57,7 @@ static int16_t find_pack(uint32_t pack_id) {
 /* ---- Slot map helpers ---- */
 
 static uint16_t slot_map_find(uint64_t resource_id) {
-    uint32_t idx = (uint32_t)(resource_id % NT_SLOT_MAP_SIZE);
+    uint32_t idx = (uint32_t)(resource_id % (uint64_t)NT_SLOT_MAP_SIZE);
     for (uint32_t i = 0; i < NT_SLOT_MAP_SIZE; i++) {
         uint32_t probe = (idx + i) % NT_SLOT_MAP_SIZE;
         uint16_t si = s_resource.slot_map[probe];
@@ -72,7 +72,7 @@ static uint16_t slot_map_find(uint64_t resource_id) {
 }
 
 static void slot_map_insert(uint64_t resource_id, uint16_t slot_index) {
-    uint32_t idx = (uint32_t)(resource_id % NT_SLOT_MAP_SIZE);
+    uint32_t idx = (uint32_t)(resource_id % (uint64_t)NT_SLOT_MAP_SIZE);
     for (uint32_t i = 0; i < NT_SLOT_MAP_SIZE; i++) {
         uint32_t probe = (idx + i) % NT_SLOT_MAP_SIZE;
         if (s_resource.slot_map[probe] == 0) {
